fix make textfield as global widget
