+++ lib/ui/screens/login/login_screens.dart
@@ -3,6 +3,7 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:susmatior_app/constants/colors_constants.dart';
 import 'package:susmatior_app/constants/padding_constants.dart';
 import 'package:susmatior_app/constants/radius_constants.dart';
+import 'package:susmatior_app/ui/screens/widgets/textfield_widget.dart';
 
 class LoginScreen extends StatelessWidget {
   const LoginScreen({Key? key}) : super(key: key);
@@ -122,31 +123,4 @@ class LoginScreen extends StatelessWidget {
   }
 }
 
-class TextFormFieldWhite extends StatelessWidget {
-  TextFormFieldWhite(
-      {Key? key,
-      TextEditingController? controller,
-      required this.label,
-      required this.isObscure})
-      : super(key: key);
 
-  String label;
-  bool isObscure;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      style: const TextStyle(color: Colors.white),
-      obscureText: isObscure,
-      cursorColor: Colors.white,
-      decoration: InputDecoration(
-        labelText: label,
-        labelStyle: const TextStyle(color: Colors.white),
-        focusedBorder: const OutlineInputBorder(
-            borderSide: BorderSide(width: 2, color: Colors.white)),
-        enabledBorder: const OutlineInputBorder(
-            borderSide: BorderSide(width: 2, color: Colors.white)),
-      ),
-    );
-  }
-}
